// src/types/enums/mqtt-commands.ts
export enum MqttCommands {
  StopEngine = 'prog0',
  StartEngine = 'prog1',
  OpenLock = 'prog4',
  CloseLock = 'prog5',
  FanOn = 'prog2',
  FanOff = 'prog3',
  Scenario6 = 'prog6',
  Scenario7 = 'prog7',
  Scenario8 = 'prog8',
  Scenario9 = 'prog9',
  IN2ControlOn = 'control2=1',
  IN2ControlOff = 'control2=0',
  ThermostatOn = 'termostat=1',
  ThermostatOff = 'termostat=0',
  StarterCrankTime1Sec = 'starter=1000',
  StarterCrankTime2And5Sec = 'starter=2500',
  StarterCrankTime4Sec = 'starter=4000',
  Update = 'update',
  WiFiOn = 'wifion',
  Reboot = 'reboot',
}